Segments をストアに移動
